Changes upgrade instructions
In the situation where you are using `docker-compose` you can use `docker-compose exec` to use the running app container. This method was already used for initialising, but in the upgrade instructions there was still references to using `docker exec`. 

The new instructions are more consistent with the initialising and a bit less complicated as you don't need to find out the container name.